remove unnecessary FindLFSFile
